requirements.txt
Django==5.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
qrcode==7.4.2
Pillow==11.0.0
WeasyPrint==60.1
python-decouple==3.8
python-dateutil==2.8.2
reportlab==4.4.1
python-barcode==0.15.1
mysql-connector-python==9.3.0
